requirements.txt
streamlit 
transformers 
torch 
PyPDF2 
python-docx